requirements.txt
Django>3
psycopg2-binary==2.8.4
cerberus-1.3.2
python-dotenv-0.12.0